[Refactoring]
