+++ lib/main.dart
@@ -4,8 +4,10 @@ import 'package:core/external/route_strings.dart';
 import 'package:flutter_game_catalogue/presentation/blocs/splash_bloc.dart';
 import 'package:flutter_game_catalogue/presentation/pages/splash_screen.dart';
 import 'package:flutter_modular/flutter_modular.dart';
+import 'package:core/core.dart';
 
 void main() {
+  Modular.init(CoreModule());
   runApp(ModularApp(
     module: AppModule(),
   ));
@@ -14,7 +16,6 @@ void main() {
 class AppModule extends MainModule {
   @override
   List<Bind> get binds => [
-        Bind((_) => RouteStrings()),
         Bind((_) => SplashBloc()),
       ];
