+++ module/core/lib/core.dart
@@ -1,10 +1,14 @@
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:core/external/image_strings.dart';
+import 'external/api_constant.dart';
+import 'external/route_strings.dart';
 
 class CoreModule extends ChildModule{
   @override
   List<Bind> get binds => [
-    Bind((_) => ImageStrings())
+    Bind((_) => ApiConstant()),
+    Bind((_) => ImageStrings()),
+    Bind((_) => RouteStrings()),
   ];
 
   @override
