+++ module/feature_game/lib/data/remote/repositories/game_repository_impl.dart
@@ -1,5 +1,5 @@
 import 'package:feature_game/data/remote/dataresources/game_api_provider.dart';
-import 'package:feature_game/domain/entities/GameResponse.dart';
+import 'package:feature_game/domain/entities/game_response.dart';
 import 'package:feature_game/domain/repositories/game_repository.dart';
 import 'package:flutter/foundation.dart';
