+++ module/feature_game/lib/domain/repositories/game_repository.dart
@@ -1,4 +1,4 @@
-import 'package:feature_game/domain/entities/GameResponse.dart';
+import 'package:feature_game/domain/entities/game_response.dart';
 
 abstract class GameRepository {
   Future<GameResponse> getGameList();
