+++ module/feature_game/lib/domain/usecases/game_use_case.dart
@@ -1,4 +1,4 @@
-import 'package:feature_game/domain/entities/GameResponse.dart';
+import 'package:feature_game/domain/entities/game_response.dart';
 import 'package:feature_game/domain/repositories/game_repository.dart';
 
 abstract class GameUseCase {
