+++ module/feature_game/lib/presentation/bloc/game_bloc.dart
@@ -1,4 +1,4 @@
-import 'package:feature_game/domain/entities/GameResponse.dart';
+import 'package:feature_game/domain/entities/game_response.dart';
 import 'package:feature_game/domain/usecases/game_use_case.dart';
 import 'package:feature_game/presentation/bloc/game_event.dart';
 import 'package:feature_game/presentation/bloc/game_state.dart';
