+++ module/feature_game/lib/presentation/bloc/game_state.dart
@@ -1,5 +1,5 @@
 import 'package:equatable/equatable.dart';
-import 'package:feature_game/domain/entities/GameResponse.dart';
+import 'package:feature_game/domain/entities/game_response.dart';
 
 abstract class GameState extends Equatable {
   const GameState();
